:sparkles: msa::clustalw autocleanup.
The function was previously hardcoded to write and delete files named "internalRsequence" with the suffixes .dnd and .aln. This lead to a race condition whilst executing more than one thread of msa::clustalw simultaneously. This was fixed on the last commit, by adding a pseudo-random hash to each temp file. This however, lead to another problem. The function did not perform the post-execution cleanup of temp files because these filenames were also hardcoded. This has been fixed